Enhance security audit reports and address critical issues in ticker implementation
- Updated security report on secure coding practices to reflect improvements in dependency management and logging practices.
- Resolved previous concerns regarding missing dependencies and logging of sensitive data.
- Improved assessment of Transport Layer Security (TLS) practices, highlighting enhancements in HTTP client architecture.
- Added additional security audit findings, confirming no critical vulnerabilities and outlining minor considerations for further hardening.
- Comprehensive analysis of spot vs derivatives tests, detailing new models and enhanced test coverage.
- Completed critical fixes in Backpack ticker implementation, aligning raw model with actual API response and ensuring proper data handling.
- Documented implementation status updates and future enhancement opportunities for ticker functionality.

+++ tests/integration/apis/backpack/account/margin_balances/positive/test_bp_margin_balances_positive.py
@@ -182,10 +182,12 @@ class TestBackpackMarginBalancesPositive:
             assert account_summary.total_position_notional >= Decimal("0")
 
             # Calculate expected total notional
-            expected_notional = sum(
-                abs(pos.size * pos.mark_price)
-                for pos in positions
-                if pos.size != 0 and pos.mark_price is not None
+            expected_notional = Decimal(
+                sum(
+                    abs(pos.size * pos.mark_price)
+                    for pos in positions
+                    if pos.size != 0 and pos.mark_price is not None
+                )
             )
 
             # Allow for small differences due to price movements
